Fixes #2: Remove dart:io import
The ini.dart file now no longer has the 'dart:io' import, allowing it to
be used in the browser or on the server. The README has been updated
with code that does file operations itself.

I thought about keeping the file methods in derivative code but it
really adds very little. If a Config::Any style loader is written then
it will be useful for that to have access to the file name to determine
the loader. Until that point it does not matter.